+++ src/packages/rte/tiptap/components/index.ts
@@ -1,1 +1,2 @@
 export * from './input-tiptap/index.js';
+export * from './toolbar/tiptap-toolbar-dropdown-base.element.js';
